[GFX] vulkan physical device selection
Added default physical device selection.

decided that the gfx library will fallback to selection 0 as in the future the gfx library will expose more information to the lib user so that they are the one who decide what the selection behaviour is instead of the lib itself.

instead the lib will give back the index to a suitable physical device for it's intended purpose.