[GFX] partial gfx refactor + added gfx database
refactoring various parts of the GFX lib to make it more maintainable and easier add assets and pipelines.
mainly abstracting things out from the gfx device into various files.

started adding GFX database which returns a index into the DB when a new object is added this index will be cached in various entities as array lookups.

some solution will be needed when loading and unloading levels. to stop indices pointing to invalid parts of the DB.

by the end of this refactor I aim not to have a fallback pipeline but a lit pipeline instead.