[GFX] continued gfx refactor to DB
TLDR: continued the fallback -> lit pipeline refactor + move to managing entities and assets via the DB.

Removed descriptor set from VulkanFallbacks struct and moved it to the DB

during the I ended up with removing the homegrown maths lib for now which is a bit of a shame, but it ended up slowing things down  a bit. I would still like to revisit this, but maybe as a dedicated project instead.

Added very simple "script" for camera controller

The simple camera controller script ended up requiring a minor refactor the WIN API input system to use RAWINPUT (WM_INPUT) from MW_MOUSEMOVE this was due to dreadful mouse stuttering when setting position and hiding the cursor.
This does mean that we use a virtual mouse for mouse delta, but we will still need a "real" cursor position for the window in the future for any UI.

Something to note here is that by using RAWINPUT API we should see sub-pixel mouse movement. I'm not totally sure how to do a 1:1 comparison of the 2 but I assume the amount of MSG calls should be a reasonable indicator.

L/R shift doesn't work with my KeyCode bindings so I've added in a default shift keycode `16`